fix: upgrade Style Dictionary
